show document card inside message
