+++ samples/apps/copilot-chat-app/webapp/src/components/chat/ChatHistory.tsx
@@ -2,9 +2,8 @@
 
 import { makeStyles, shorthands, tokens } from '@fluentui/react-components';
 import React from 'react';
-import { ChatMessageType, IChatMessage } from '../../libs/models/ChatMessage';
+import { IChatMessage } from '../../libs/models/ChatMessage';
 import { GetResponseOptions } from '../../libs/useChat';
-import { ChatHistoryFileItem } from './ChatHistoryFileItem';
 import { ChatHistoryItem } from './ChatHistoryItem';
 
 const useClasses = makeStyles({
@@ -35,18 +34,12 @@ export const ChatHistory: React.FC<ChatHistoryProps> = ({ messages, onGetRespons
             {messages
                 .slice()
                 .sort((a, b) => a.timestamp - b.timestamp)
-                .map((message, index) =>
-                    message.type === ChatMessageType.Document ? (
-                        <ChatHistoryFileItem key={message.timestamp} message={message} />
-                    ) : (
-                        <ChatHistoryItem
-                            key={message.timestamp}
-                            message={message}
-                            getResponse={onGetResponse}
-                            messageIndex={index}
-                        />
-                    ),
-                )}
+                .map((message, index) => <ChatHistoryItem
+                    key={message.timestamp}
+                    message={message}
+                    getResponse={onGetResponse}
+                    messageIndex={index}
+                />)}
         </div>
     );
 };
